Refactor folder structure and fix eslint
