+++ .gitignore
@@ -3,10 +3,12 @@ db_backups
 
 # User-specific stuff
 .idea/**/workspace.xml
+.idea/**/misc.xml
 .idea/**/tasks.xml
 .idea/**/usage.statistics.xml
 .idea/**/dictionaries
 .idea/**/shelf
+.idea/**/codeStyles
 
 # AWS User-specific
 .idea/**/aws.xml
