+++ client/eslint.config.mjs
@@ -1,0 +1,95 @@
+import { fixupConfigRules, fixupPluginRules } from "@eslint/compat";
+import typescriptEslint from "@typescript-eslint/eslint-plugin";
+import react from "eslint-plugin-react";
+import reactHooks from "eslint-plugin-react-hooks";
+import globals from "globals";
+import tsParser from "@typescript-eslint/parser";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import js from "@eslint/js";
+import { FlatCompat } from "@eslint/eslintrc";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const compat = new FlatCompat({
+    baseDirectory: __dirname,
+    recommendedConfig: js.configs.recommended,
+    allConfig: js.configs.all
+});
+
+export default [...fixupConfigRules(compat.extends(
+    "plugin:react/recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier",
+    "plugin:prettier/recommended"
+)), {
+    files: [
+      "src/**/*.js",
+      "src/**/*.jsx",
+      "src/**/*.ts",
+      "src/**/*.tsx",
+    ],
+
+    plugins: {
+        "@typescript-eslint": fixupPluginRules(typescriptEslint),
+        react: fixupPluginRules(react),
+        "react-hooks": fixupPluginRules(reactHooks),
+    },
+
+    languageOptions: {
+        globals: {
+            ...globals.browser,
+            ...globals.jest,
+        },
+
+        parser: tsParser,
+        ecmaVersion: "latest",
+        sourceType: "module",
+
+        parserOptions: {
+            ecmaFeatures: {
+                jsx: true,
+            },
+
+            project: ["tsconfig.json"],
+        },
+    },
+
+    settings: {
+        react: {
+            version: "detect",
+        },
+    },
+
+    rules: {
+        "no-shadow": "off",
+        "@typescript-eslint/no-shadow": ["error"],
+
+        "@typescript-eslint/explicit-function-return-type": ["error", {
+            allowExpressions: true,
+        }],
+
+        "@typescript-eslint/no-explicit-any": "off",
+        "@typescript-eslint/strict-boolean-expressions": "off",
+        "@typescript-eslint/return-await": "off",
+
+        "max-len": ["warn", {
+            code: 160,
+            ignoreComments: true,
+            ignoreUrls: true,
+        }],
+
+        "react/react-in-jsx-scope": "off",
+
+        "react/jsx-filename-extension": [1, {
+            extensions: [".tsx", ".jsx"],
+        }],
+
+        "react-hooks/rules-of-hooks": "error",
+        "react-hooks/exhaustive-deps": "warn",
+
+        "prettier/prettier": ["error", {
+            endOfLine: "auto",
+        }]
+    },
+}];
