+++ client/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "client",
+  "name": "thesis-track",
   "version": "2.0.0",
   "main": "src/index.tsx",
   "license": "MIT",
   "scripts": {
     "dev": "webpack serve --open --mode development",
-    "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
-    "lint:fix": "eslint --fix \"src/**/*.{js,jsx,ts,tsx}\"",
+    "lint": "eslint src/**/*.{js,jsx,ts,tsx}",
+    "lint:fix": "eslint --fix src/**/*.{js,jsx,ts,tsx}",
     "build": "webpack --mode production"
   },
   "dependencies": {
@@ -53,6 +53,9 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.1.0",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.5.0",
     "@tanstack/eslint-plugin-query": "^5.35.6",
     "@types/copy-webpack-plugin": "^10.1.0",
     "@types/dotenv-webpack": "^7.0.7",
@@ -66,11 +69,11 @@
     "dotenv-webpack": "^8.1.0",
     "eslint": "^9.3.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^17.7.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-react-hooks": "^4.6.2",
+    "globals": "^15.6.0",
     "html-webpack-plugin": "^5.6.0",
     "prettier": "^3.2.5",
     "sass": "^1.75.0",
@@ -79,5 +82,8 @@
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4"
   },
-  "packageManager": "yarn@4.3.1"
+  "packageManager": "yarn@4.3.1",
+  "engines": {
+    "node": ">=18.0.0"
+  }
 }
