+++ client/src/interfaces/authentication.ts
