+++ client/src/pages/ThesisApplication/components/ApplicationSubmission/ApplicationSuccessfulSubmission.tsx
