+++ client/src/pages/ThesisApplication/components/DeclarationOfDataConsent/DeclarationOfDataConsent.tsx
