+++ client/src/pages/ThesisOverview/ThesisOverview.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const ThesisOverview = (): JSX.Element => {
+  return <div style={{ padding: '0 0 5vh 0' }}> </div>
+}
+
+export default ThesisOverview
