+++ client/src/pages/ThesisOverview/components/ThesisProgressChart/ThesisProgressChart.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const ThesisProgressChart = (): JSX.Element => {
+  return <div style={{ padding: '0 0 5vh 0' }}></div>
+}
+
+export default ThesisProgressChart
