+++ client/tsconfig.json
@@ -11,7 +11,7 @@
     "moduleResolution": "node",
     "strict": true,
     "skipLibCheck": true,
-    "noImplicitAny": false
+    "noImplicitAny": false,
   },
   "include": ["**/*.ts", "**/*.tsx"]
 }
