+++ client/webpack.config.ts
@@ -1,6 +1,5 @@
 import path from 'path'
 import { Configuration, EnvironmentPlugin } from 'webpack'
-// eslint-disable-next-line import/default
 import CopyPlugin from 'copy-webpack-plugin'
 import 'webpack-dev-server'
 import HtmlWebpackPlugin from 'html-webpack-plugin'
